materialized: push simple sql execution into coord client
This will make it possible to use the simple SQL execution interface
outside of the HTTP client. Also I think it's just nice to have it as
part of the coordinator's API.
