+++ src/materialized/Cargo.toml
@@ -70,7 +70,6 @@ serde = { version = "1.0.124", features = ["derive"] }
 serde_json = "1.0.64"
 shell-words = "1.0.0"
 sql = { path = "../sql" }
-sql-parser = { path = "../sql-parser" }
 structopt = "0.3.21"
 sysctl = "0.4.0"
 sysinfo = "0.16.4"
